strip dataset features from base as this is now all handled in holder
bundle
